+++ src/web/worker-bridge-server.js
@@ -16,6 +16,7 @@ var BridgeServer = require('./bridge-server.js');
 // - `config.temp`: optional bool, instructs the worker to self-destruct after its finished responding to its requests
 // - `config.log`: optional bool, enables logging of all message traffic
 function WorkerBridgeServer(config) {
+	var self = this;
 	if (!config || (!config.src && !config.domain))
 		throw new Error("WorkerBridgeServer requires config with `src` or `domain` attribute.");
 	BridgeServer.call(this, config);
@@ -54,12 +55,15 @@ function WorkerBridgeServer(config) {
 					var script_blob = new Blob([bootstrap_src+'(function(){'+res.body+'})();'], { type: "text/javascript" });
 					config.src = window.URL.createObjectURL(script_blob);
 					src_.fulfill(config.src);
+				})
+				.fail(function(res) {
+					src_.reject(null);
+					self.terminate(404, 'Worker Not Found');
 				});
 		}
 	}
 
-	var self = this;
-	src_.always(function() {
+	src_.then(function() {
 		// Prep config
 		if (!self.config.domain) { // assign a temporary label for logging if no domain is given yet
 			self.config.domain = '<'+self.config.src.slice(0,40)+'>';
@@ -110,9 +114,9 @@ WorkerBridgeServer.prototype.getPort = function() {
 	return this.worker.port ? this.worker.port : this.worker;
 };
 
-WorkerBridgeServer.prototype.terminate = function() {
-	BridgeServer.prototype.terminate.call(this);
-	this.worker.terminate();
+WorkerBridgeServer.prototype.terminate = function(status, reason) {
+	BridgeServer.prototype.terminate.call(this, status, reason);
+	if (this.worker) this.worker.terminate();
 	this.worker = null;
 	this.isActive = false;
 };
